feat: add verify update method
